Updated to using 24 bit counter

+++ CSynth/synth.h
@@ -2,8 +2,8 @@
 #include "globals.h"
 
 
-extern u16 WC;
-extern u16 INCR;
+extern u32 WC;
+extern u32 INCR;
 
 void Tick();
